deploy/pythonanywhere/requirements.txt
-f /usr/share/pip-wheels
alembic==0.9.9
asn1crypto==0.24.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
blinker==1.4
bokeh==0.12.14
certifi==2018.1.18
cffi==1.11.5
click==6.7
cryptography==2.2.2
dill==0.2.6
enum34==1.1.6
Flask==0.12.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-User==1.0.1.3
Flask-WTF==0.14.2
futures==3.2.0
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
ndg-httpsclient==0.4.4
numpy==1.14.2
oauth==1.0.1
oauthlib==2.0.2
ordereddict==1.1
packaging==17.1
pandas==0.20.3
passlib==1.7.1
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.3
PyYAML==3.12
requests==2.7.0
scipy==1.0.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.7
tornado==4.3
untangle==1.1.1
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
